chore: split ataxx-uai into files
